requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
pydantic-settings
python-dotenv
yfinance
pandas